scripts/tee/requirements.txt
PyYAML==3.10
